Time: 49 ms (19.58%), Space: 27.4 MB (99.88%) - LeetHub

+++ 0704-binary-search/0704-binary-search.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int search(vector<int>& nums, int target) {
         
-        long long start=0,end=nums.size()-1,mid=(end+start)/2;
+        int start=0,end=nums.size()-1,mid=(end+start)/2;
         
         while(start<=end){
             mid=(end+start)/2;
